refactor: extract splitLargeChunk to shared utility
- Create chunk-utils.ts with BaseChunk interface and splitLargeChunk function
- Both ASTChunk and TreeSitterChunk now alias BaseChunk
- Reduces code duplication between ast-chunker and tree-sitter-chunker
